+++ orion/sites/admin/src/announce/announce.js
@@ -1,6 +1,7 @@
 "use strict";
 require("./announce.styl");
 import React from "react";
+import moment from "moment";
 import API from "../api.js";
 import { Link } from "react-router-dom";
 
@@ -33,6 +34,7 @@ export class AnnouncePage extends React.Component {
                 <h1>All Announcements ({numRows}) </h1>
 
                 <ul className="announce-list-row subheader">
+                    <li className="li-med">State</li>
                     <li className="li-med">Date</li>
                     <li className="li-med">Author</li>
                     <li className="li-large">Message</li>
@@ -52,18 +54,27 @@ export class AnnouncePage extends React.Component {
 class AnnounceRow extends React.Component {
     render() {
         const announceId = this.props.row.id;
-        const date = new Date(this.props.row.postedAt);
-        const dateLine1 = date.toDateString();
-        const dateLine2 = date.toLocaleTimeString();
+        const postedAt = moment(this.props.row.postedAt);
+
+        const now = moment();
+        const isScheduled = postedAt.isAfter(now);
+
         const author = this.props.row.author;
         const message = this.props.row.message;
 
         const url = "/announcements/" + announceId + "/edit";
         return (
             <ul className="announce-list-row">
+                <li
+                    className={
+                        "li-med " + (isScheduled ? " scheduled" : " published")
+                    }>
+                    <div>{isScheduled ? "Scheduled" : "Published"}</div>
+                    <div>{postedAt.fromNow()}</div>
+                </li>
                 <li className="li-med">
-                    <div>{dateLine1}</div>
-                    <div>{dateLine2}</div>
+                    <div>{postedAt.format("M/D/YYYY")}</div>
+                    <div>{postedAt.format("hh:mm a")}</div>
                 </li>
                 <li className="li-med"> {author} </li>
                 <li className="li-large"> {message} </li>
