+++ orion/sites/admin/src/announce/announce.styl
@@ -33,8 +33,16 @@
 
         &.li-med
           width: spacing24x
+
+          &.scheduled
+            font-style: italic
+            color: colorGrayMed
+          &.published
+            color: colorTurquoise
+
         &.li-large
-          width: spacing48x
+          width: spacing56x
+          margin-right: spacing4x
 
         & div
           line-height: lineHeight1
