+++ orion/sites/admin/src/announce/announceEdit.js
@@ -1,16 +1,9 @@
 "use strict";
 require("./announceEdit.styl");
 import React from "react";
-
 import moment from "moment";
-import "react-dates/initialize";
-import { SingleDatePicker } from "react-dates";
-import "react-dates/lib/css/_datepicker.css";
-
-import TimePicker from "react-times";
-import "react-times/css/classic/default.css";
-
 import API from "../api.js";
+import { AnnounceDateTimeSection } from "./announceEditDateTime.js";
 import { Modal } from "../modals/modal.js";
 import { OkayModal } from "../modals/okayModal.js";
 import { YesNoModal } from "../modals/yesnoModal.js";
@@ -21,7 +14,7 @@ export class AnnounceEditPage extends React.Component {
         this.state = {
             announceId: 0,
             datePickerFocused: false,
-            inputPostedAt: moment(), // always a moment object
+            inputPostedAt: null,
             inputAuthor: "",
             inputMessage: "",
             isEdit: false,
@@ -171,7 +164,7 @@ export class AnnounceEditPage extends React.Component {
                 {modalDiv}
                 <h2>{title}</h2>
 
-                <AnnounceDateTimePicker
+                <AnnounceDateTimeSection
                     postedAt={this.state.inputPostedAt}
                     onMomentChange={this.onMomentChange}
                 />
@@ -201,109 +194,3 @@ export class AnnounceEditPage extends React.Component {
         );
     }
 }
-
-class AnnounceDateTimePicker extends React.Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            dateFocused: false
-        };
-        this.onClickNow = this.onClickNow.bind(this);
-        this.onDateChange = this.onDateChange.bind(this);
-    }
-
-    onClickNow() {
-        const now = moment();
-        this.props.onMomentChange(now);
-    }
-
-    onDateChange(date) {
-        let dateMoment = moment(date);
-        let newMoment = this.props.postedAt
-                            .year(dateMoment.year())
-                            .month(dateMoment.month())
-                            .date(dateMoment.date());
-        this.props.onMomentChange(newMoment);
-    }
-
-    onTimeChange(options) {
-        let newHour = parseInt(options.hour);
-        let hour = options.meridiem == "AM" ? newHour : newHour + 12;
-        let minute = parseInt(options.minute);
-        let newMoment = this.props.postedAt
-                            .hour(hour)
-                            .minute(minute);
-        this.props.onMomentChange(newMoment);
-    }
-
-    onFocusChange(focusState) {
-        // For TimePicker. Do nothing
-    }
-
-    render() {
-        const postedAt = this.props.postedAt;
-        const postedAtString = postedAt.format("dddd, MMMM Do YYYY, h:mm az").toString();
-
-        return (
-            <div id="announce-datetime-container">
-                <div id="announce-date-picker">
-                    <h4>Select Date</h4>
-                    <SingleDatePicker
-                        date={postedAt}
-                        onDateChange={(date) => this.onDateChange(date)}
-                        focused={this.state.dateFocused}
-                        onFocusChange={({ focused }) =>
-                            this.setState({ dateFocused: focused })
-                        }
-                        id="announce-date-picker"
-                        showDefaultInputIcon
-                    />
-                </div>
-
-                <div id="announce-time-picker">
-                    <h4>Select Time</h4>
-                    <TimePicker
-                        time={postedAt.format("HH:mm")}
-                        theme="classic"
-                        showTimezone={true}
-                        timeMode="12"
-                        onFocusChange={this.onFocusChange.bind(this)}
-                        onTimeChange={this.onTimeChange.bind(this)}
-                        minuteStep={15}
-                        timeConfig={{
-                            from: "7:00 AM",
-                            to: "11:45 PM",
-                            step: 15,
-                            unit: "minutes",
-                        }}
-                    />
-                </div>
-                <p>Or</p>
-                <button className="btn-now" onClick={this.onClickNow}>Schedule For Now</button>
-
-                <h4>Announcement will be published on <b>{postedAtString}</b></h4>
-            </div>
-        );
-    }
-}
-
-
-// console.log("**************************");
-// console.log(postedAt.toString());
-// console.log(postedAt.toISOString());
-// console.log(postedAt.toJSON());
-// console.log(postedAt.toObject());
-
-// console.log(postedAt.toDate());
-// console.log(postedAt.toDate().toLocaleString());
-// console.log(postedAt.toDate().toJSON());
-// console.log("**************************");
-// **************************
-// Thu Apr 09 2020 12:00:00 GMT-0400
-// 2020-04-09T16:00:00.000Z
-// 2020-04-09T16:00:00.000Z
-// {years: 2020, months: 3, date: 9, hours: 12, minutes: 0, …}years: 2020months: 3date: 9hours: 12minutes: 0seconds: 0milliseconds: 0__proto__: Object
-// Thu Apr 09 2020 12:00:00 GMT-0400 (Eastern Daylight Time)
-// 4/9/2020, 12:00:00 PM
-// 2020-04-09T16:00:00.000Z
-// **************************
