+++ orion/sites/admin/src/static.styl
@@ -7,6 +7,7 @@ colorWhiteLight = #f2f2f2
 colorBlackDark = #222
 colorBlack = #000000
 colorGrayDark = #414141
+colorGrayMed = #787878
 colorGrayLight = #b8b8b8
 
 // Fonts - Font Families
@@ -49,6 +50,7 @@ spacing32x = (spacing1x * 32)
 spacing36x = (spacing1x * 36)
 spacing42x = (spacing1x * 42)
 spacing48x = (spacing1x * 48)
+spacing56x = (spacing1x * 56)
 spacing64x = (spacing1x * 64)
 spacing96x = (spacing1x * 96)
 spacing128x = (spacing1x * 128)
